Split motor controller from bridge module
